+++ index.js
@@ -6,13 +6,14 @@ const prompt = require("./src/prompt");
 const intCmds = require("./src/interpret");
 
 const args = _mainArgs();
+
 if (args.length !== 0) {
   const command = args.filter((v) => {
     if (v.startsWith("-") || v.startsWith("/")) return false;
     return true;
   });
 
-  intCmds(command.join(" "), false);
+  intCmds(command.join(" "));
   process.exit(0);
 } else {
   require("./src/intro");
@@ -20,8 +21,6 @@ if (args.length !== 0) {
   // Repeat until the user exits
   while (true) {
     // Ask the user for a command
-    const { command, isEmpty } = prompt();
-
-    intCmds(command, isEmpty);
+    intCmds(prompt());
   }
 }